Fixed MySQL port type

+++ php/config.php
@@ -4,7 +4,7 @@ $host = getenv("MYSQLHOST");
 $user = getenv("MYSQLUSER");
 $password = getenv("MYSQLPASSWORD");
 $database = getenv("MYSQLDATABASE");
-$port = getenv("MYSQLPORT");
+$port = (int)getenv("MYSQLPORT");
 
 $conn = new mysqli(
     $host,
